extracted Addressables Duplication Tests
